V4.2 develop

+++ AiyaEffectSDK/AyFaceTrack/src/main/java/com/aiyaapp/aiya/AyFaceTrack.java
@@ -16,19 +16,25 @@ public class AyFaceTrack {
         System.loadLibrary("AyFaceTrackJni");
     }
 
+    private static boolean isCopiedAssetResource = false;
     /**
      * 初始化人脸识别
      */
     public static void Init(Context context) {
-        File folder = context.getExternalCacheDir();
+        File folder = context.getFilesDir();
         if (folder == null) {
             folder = context.getCacheDir();
         }
 
         if (folder != null) {
+
             String dstPath = folder.getAbsolutePath() + "/aiya/config";
-            deleteFile(new File(dstPath));
-            copyFileFromAssets("config", dstPath, context.getAssets());
+
+            if (!isCopiedAssetResource) { // 只拷贝一次
+                deleteFile(new File(dstPath));
+                copyFileFromAssets("config", dstPath, context.getAssets());
+                isCopiedAssetResource = true;
+            }
 
             Init(dstPath);
         }
@@ -37,7 +43,7 @@ public class AyFaceTrack {
     private static native void Init(String dstPath);
     public static native void Deinit();
     public static native long FaceData();
-    public static native void TrackWithBGRABuffer(ByteBuffer pixelBuffer, int width, int height);
+    public static native int TrackWithBGRABuffer(ByteBuffer pixelBuffer, int width, int height);
 
     public static void deleteFile(File file) {
         if (file.exists()) {
